### Description
<!-- Describe your changes. -->

Update SQNBitGemm ARM NEON kernel to compute 4x2 tile of output.

Note: Also tried 2x4 and 4x4 tiles but observed the best microbenchmark results with 4x2 tiles.

#### Measurements

Baseline: 20cd3394fc
Updated: aecb18a3c7

##### Microbenchmarks

Run on Azure VM (ARM64 Linux) with compute type: CompInt8, number of threads: 4, M:128/K:4096/N:4096

blklen|symmetric|baseline time (ns)|updated time (ns)
-|-|-|-
16 |1 |51511766|44803227
16 |0 |58870228|49002040
32 |1 |29887367|25812083
32 |0 |33208816|26632430
64 |1 |30344972|26624130
64 |0 |31460702|25966747

##### E2E test

Run onnxruntime-genai benchmark with Phi-3 mini using 4 threads.

machine|baseline prompt processing tokens/second|updated pp t/s
-|-|-
Samsung Galaxy S21|15.93|17.93
Surface Pro 9|33.36|36.60
Azure VM|18.92|21.84

### Motivation and Context
<!-- - Why is this change required? What problem does it solve?
- If it fixes an open issue, please link to the issue here. -->

Improve prompt processing (M>1) performance.